State tests working

+++ tp2_utils_package/test_integration_keep_state.py
@@ -13,70 +13,51 @@ from tp2_utils.message_pipeline.operations.operation import Operation
 from tp2_utils.message_pipeline.message_set.disk_message_set import DiskMessageSet
 from tp2_utils.rabbit_utils.publisher_sharding import PublisherSharding
 from tp2_utils.rabbit_utils.rabbit_consumer_producer import RabbitQueueConsumerProducer
+import random
+from time import sleep
+import numpy as np
+import signal
 
-DEFAULT_MESSAGES_TO_GROUP = 5
+DEFAULT_MESSAGES_TO_GROUP = 1000
 
-OPERATION_ARGS = [[]]
+OPERATION_ARGS = [([], []),
+                  ([["GroupBy", "key"]], [[{'name': "Count"}]]),
+                  ([["GroupBy", "key"]], [[{'name': "Count"}]]),
+                  ([["GroupBy", "key"]], [[{'name': "Count"}]]),
+                  ([["Filter", "count", lambda x: x > 2]], [[]])
+                  ]
 
-PIPELINE_C_ARGS = [({'ends_to_receive':1, 'ends_to_send':1},
+PIPELINE_C_ARGS = [({'ends_to_receive':1, 'ends_to_send':1, 'data_path': '/tmp/datapath1'},
                     {'host': "localhost", 'consume_queue': 'pipeline_start',
                      'response_queues': ['pipelineC_step1'],
                      'messages_to_group': DEFAULT_MESSAGES_TO_GROUP},
                     ['/tmp/message_set1'],
-                    ['key',3])]
-
-MESSAGES_FOR_TESTING_ENDING_1 = [
-    {'key': 'A', '_id': 0},
-    {'key': 'A', '_id': 0},
-    {'key': 'A', '_id': 1},
-    [{'key': 'B', '_id': 2}, {'key': 'C', '_id': 3}],
-    [{'key': 'B', '_id': 4}, {'key': 'D', '_id': 5}],
-    [{'key': 'D', '_id': 6}, {'key': 'C', '_id': 7}],
-    [{'key': 'B', '_id': 8}, {'key': 'C', '_id': 9}],
-    [{'key': 'B', '_id': 10}, {'key': 'D', '_id': 11}],
-    [{'key': 'D', '_id': 12}, {'key': 'C', '_id': 13}],
-    [{'key': 'B', '_id': 14}, {'key': 'C', '_id': 15}],
-    [{'key': 'B', '_id': 16}, {'key': 'D', '_id': 17}],
-    [{'key': 'D', '_id': 18}, {'key': 'C', '_id': 19}],
-    [{'key': 'E', '_id': 20}, {'key': 'E', '_id': 21}],
-    {'key': 'F', '_id': 22},
-    {'key': 'F', '_id': 23},
-    {'key': 'G', '_id': 24},
-    {'key': 'G', '_id': 25},
-    {'key': 'G', '_id': 26},
-    [{'key': 'H', '_id': 26}, {'key': 'H', '_id': 27}, {'key': 'H', '_id': 28},
-     {'key': 'H', '_id': 29}, {'key': 'H', '_id': 30}, {'key': 'H', '_id': 31}],
-    [{'key': 'I', '_id': 32}, {'key': 'J', '_id': 33}, {'key': 'K', '_id': 34},
-     {'key': 'L', '_id': 35}, {'key': 'M', '_id': 36}],
-    WINDOW_END_MESSAGE
-]
-
-MESSAGES_FOR_TESTING_ENDING_2 = [
-    {'key': 'A', '_id': 0},
-    {'key': 'A', '_id': 0},
-    {'key': 'A', '_id': 1},
-    [{'key': 'B', '_id': 2}, {'key': 'C', '_id': 3}],
-    [{'key': 'B', '_id': 4}, {'key': 'D', '_id': 5}],
-    [{'key': 'D', '_id': 6}, {'key': 'C', '_id': 7}],
-    [{'key': 'B', '_id': 8}, {'key': 'C', '_id': 9}],
-    [{'key': 'B', '_id': 10}, {'key': 'D', '_id': 11}],
-    [{'key': 'D', '_id': 12}, {'key': 'C', '_id': 13}],
-    [{'key': 'B', '_id': 14}, {'key': 'C', '_id': 15}],
-    [{'key': 'B', '_id': 16}, {'key': 'D', '_id': 17}],
-    [{'key': 'D', '_id': 18}, {'key': 'C', '_id': 19}],
-    [{'key': 'E', '_id': 20}, {'key': 'E', '_id': 21}],
-    {'key': 'F', '_id': 22},
-    {'key': 'F', '_id': 23},
-    {'key': 'G', '_id': 24},
-    {'key': 'G', '_id': 25},
-    {'key': 'G', '_id': 26},
-    [{'key': 'H', '_id': 26}, {'key': 'H', '_id': 27}, {'key': 'H', '_id': 28},
-     {'key': 'H', '_id': 29}, {'key': 'H', '_id': 30}, {'key': 'H', '_id': 31}],
-    [{'key': 'I', '_id': 32}, {'key': 'J', '_id': 33}, {'key': 'K', '_id': 34},
-     {'key': 'L', '_id': 35}, {'key': 'M', '_id': 36}, WINDOW_END_MESSAGE]
-]
-
-TESTING_RESULT = {'B': 6, 'C': 6, 'D': 6, 'G': 3, 'H': 6}
+                    ['key',3]),
+                   ({'ends_to_receive': 1, 'ends_to_send': 1, 'data_path': '/tmp/datapath2'},
+                    {'host': "localhost", 'consume_queue': 'pipelineC_step1_shard0',
+                     'response_queues': ['pipelineC_step2'],
+                     'messages_to_group': DEFAULT_MESSAGES_TO_GROUP},
+                    ['/tmp/message_set2'],
+                    None),
+                    ({'ends_to_receive': 1, 'ends_to_send': 1, 'data_path': '/tmp/datapath3'},
+                    {'host': "localhost", 'consume_queue': 'pipelineC_step1_shard1',
+                     'response_queues': ['pipelineC_step2'],
+                     'messages_to_group': DEFAULT_MESSAGES_TO_GROUP},
+                    ['/tmp/message_set3'],
+                    None),
+                   ({'ends_to_receive': 1, 'ends_to_send': 1, 'data_path': '/tmp/datapath4'},
+                    {'host': "localhost", 'consume_queue': 'pipelineC_step1_shard2',
+                     'response_queues': ['pipelineC_step2'],
+                     'messages_to_group': DEFAULT_MESSAGES_TO_GROUP},
+                    ['/tmp/message_set4'],
+                    None),
+                   ({'ends_to_receive': 3, 'ends_to_send': 1, 'data_path': '/tmp/datapath5'},
+                    {'host': "localhost", 'consume_queue': 'pipelineC_step2',
+                     'response_queues': ['pipelineC_result'],
+                     'messages_to_group': DEFAULT_MESSAGES_TO_GROUP},
+                    ['/tmp/message_set5'],
+                    None)
+                   ]
 
 class TestIntegrations(unittest.TestCase):
     """
@@ -92,17 +73,22 @@ class TestIntegrations(unittest.TestCase):
         self.channel.queue_purge(queue_name)
         self.queues_to_purge.append(queue_name)
 
-    def _setup_message_set(self, location: str):
-        shutil.rmtree(location, ignore_errors=True)
-        os.mkdir(location)
+    def _self_register_dir(self, location, delete_stuff=True):
+        if delete_stuff:
+            shutil.rmtree(location, ignore_errors=True)
+            os.mkdir(location)
+        if location not in self.dirs_to_delete:
+            self.dirs_to_delete.append(location)
+
+    def _setup_message_set(self, location: str, delete_stuff=True):
+        self._self_register_dir(location, delete_stuff)
         message_set = DiskMessageSet(location)
-        self.dirs_to_delete.append(location)
         return message_set
 
-    def _setup_start_process(self, func):
+    def _setup_start_process(self, func, ops_args, process_args):
         p = Process(target=func)
         p.start()
-        self.processes_to_join.append(p)
+        self.processes_to_join.append((p, ops_args, process_args))
 
     @staticmethod
     def _read_process(write_pipe: Pipe, final_queue: str):
@@ -120,22 +106,42 @@ class TestIntegrations(unittest.TestCase):
                               on_message_callback=partial(consume, write_pipe))
         channel.start_consuming()
 
-    def _factory_and_start_process(self, pipe_kwargs, cp_kwargs,
-                                   message_set_args=None, sharding_args=None,
-                                   operations=None):
+    def _factory_operation_list(self, ops_args):
+        final_operations = []
+        for ops, aggs in zip(ops_args[0], ops_args[1]):
+            aggs_objs = []
+            if ops[0] == "GroupBy":
+                for ag in aggs:
+                    aggs_objs.append(GroupAggregate.factory(**ag))
+            if aggs_objs:
+                final_operations.append(Operation.factory(*ops, aggregates=aggs_objs))
+            else:
+                final_operations.append(Operation.factory(*ops))
+        return final_operations
+
+    def _factory_cp(self, ops_args, process_args, delete_stuff=True):
+        pipe_kwargs, cp_kwargs, message_set_args, sharding_args = process_args
+        operations = self._factory_operation_list(ops_args)
         message_set = None
         if message_set_args:
-            message_set = self._setup_message_set(*message_set_args)
+            message_set = self._setup_message_set(*message_set_args, delete_stuff=delete_stuff)
         p_sharding = None
         if sharding_args:
             p_sharding = PublisherSharding(*sharding_args)
+        if 'data_path' in pipe_kwargs:
+            self._self_register_dir(pipe_kwargs['data_path'], delete_stuff)
         pipe = MessagePipeline(**pipe_kwargs,
                                operations=(operations if operations else []),
                                idempotency_set=message_set)
         cp = RabbitQueueConsumerProducer(**cp_kwargs,
                                          callable_commiter=pipe,
                                          publisher_sharding=p_sharding)
-        self._setup_start_process(cp)
+        return cp
+
+
+    def _factory_and_start_process(self, ops_args, process_args):
+        self._setup_start_process(self._factory_cp(ops_args, process_args),
+                                  ops_args, process_args)
 
     def _setup_pipelineC(self):
         self._setup_queue('pipelineC_step1_shard0')
@@ -144,65 +150,42 @@ class TestIntegrations(unittest.TestCase):
         self._setup_queue('pipelineC_step2')
         self._setup_queue('pipelineC_result')
         for ops_args, process_args in zip(OPERATION_ARGS, PIPELINE_C_ARGS):
-            self._factory_and_start_process(*process_args)
+            self._factory_and_start_process(ops_args, process_args)
 
-        # receive sharded messages and count
+    def _chaos_monkey_process(self, write_chaos):
+        chaos_count = 0
+        while True:
+            chaos_count += 1
+            i = random.choice(list(range(len(self.processes_to_join))))
+            self.processes_to_join[i][0].terminate()
+            cp = self._factory_cp(self.processes_to_join[i][1],
+                                  self.processes_to_join[i][2],
+                                  delete_stuff=False)
+            p = Process(target=cp)
+            p.start()
+            write_chaos.send(p.pid)
+            self.processes_to_join[i]=(p,self.processes_to_join[i][1], self.processes_to_join[i][2])
+            sleep(np.random.poisson(2.0, size=1)[0])
 
-        message_set = self._setup_message_set('/tmp/message_set2')
-        pipe = MessagePipeline([Operation.factory("GroupBy", group_by="key", aggregates=[
-            GroupAggregate.factory("Count"),
-        ])],
-                               ends_to_receive=1, ends_to_send=1,
-                               idempotency_set=message_set)
-        cp = RabbitQueueConsumerProducer("localhost", 'pipelineC_step1_shard0',
-                                         ['pipelineC_step2'],
-                                         pipe,
-                                         messages_to_group=DEFAULT_MESSAGES_TO_GROUP)
-        self._setup_start_process(cp)
-
-        message_set = self._setup_message_set('/tmp/message_set3')
-        pipe = MessagePipeline([Operation.factory("GroupBy", group_by="key", aggregates=[
-            GroupAggregate.factory("Count"),
-        ])],
-                               ends_to_receive=1, ends_to_send=1,
-                               idempotency_set=message_set)
-        cp = RabbitQueueConsumerProducer("localhost", 'pipelineC_step1_shard1',
-                                         ['pipelineC_step2'],
-                                         pipe,
-                                         messages_to_group=DEFAULT_MESSAGES_TO_GROUP)
-        self._setup_start_process(cp)
-
-        message_set = self._setup_message_set('/tmp/message_set4')
-        pipe = MessagePipeline([Operation.factory("GroupBy", group_by="key", aggregates=[
-            GroupAggregate.factory("Count"),
-        ])],
-                               ends_to_receive=1, ends_to_send=1,
-                               idempotency_set=message_set)
-        cp = RabbitQueueConsumerProducer("localhost", 'pipelineC_step1_shard2',
-                                         ['pipelineC_step2'],
-                                         pipe,
-                                         messages_to_group=DEFAULT_MESSAGES_TO_GROUP)
-        self._setup_start_process(cp)
-
-        # filter
-        message_set = self._setup_message_set('/tmp/message_set5')
-        pipe = MessagePipeline([Operation.factory("Filter", "count", lambda x: x > 2)],
-                               ends_to_receive=3, ends_to_send=1,
-                               idempotency_set=message_set)
-        cp = RabbitQueueConsumerProducer("localhost", 'pipelineC_step2',
-                                         ['pipelineC_result'],
-                                         pipe,
-                                         messages_to_group=DEFAULT_MESSAGES_TO_GROUP, )
-        self._setup_start_process(cp)
-
-    def test_count_pipeline_C_setup_ok(self):
+    def test_without_chaos_monkey(self):
+        expected_count = {}
+        for i in range(100):
+            list_of_elements = []
+            for j in range(1000):
+                element = {'key': chr(ord('A') + random.randint(0,25)), '_id': i*1000+j}
+                if element['key'] not in expected_count:
+                    expected_count[element['key']]=1
+                else:
+                    expected_count[element['key']] += 1
+                list_of_elements.append(element)
+            self.channel.basic_publish(exchange='', routing_key='pipeline_start',
+                                       body=json.dumps(list_of_elements))
+        self.channel.basic_publish(exchange='', routing_key='pipeline_start',
+                                   body=json.dumps({}))
         self._setup_pipelineC()
         consume_process = Process(target=self._read_process, args=(self.write_pipe,
                                                                    'pipelineC_result'))
         consume_process.start()
-        for element in MESSAGES_FOR_TESTING_ENDING_1:
-            self.channel.basic_publish(exchange='', routing_key='pipeline_start',
-                                       body=json.dumps(element))
         consume_process.join()
         processed_data = []
         while not processed_data or processed_data[-1] != WINDOW_END_MESSAGE:
@@ -216,14 +199,29 @@ class TestIntegrations(unittest.TestCase):
                     count_result[item['key']] = item['count']
             else:
                 count_result[resp['key']] = resp['count']
-        self.assertEqual(TESTING_RESULT, count_result)
+        self.assertEqual({k:v for k,v in  expected_count.items() if v>2}, count_result)
 
+    def test_with_chaos_monkey(self):
+        self._setup_pipelineC()
+        chaos_monkey_p = Process(target=self._chaos_monkey_process, args=(self.write_chaos,))
+        chaos_monkey_p.start()
+        expected_count = {}
+        for i in range(100):
+            list_of_elements = []
+            for j in range(1000):
+                element = {'key': chr(ord('A') + random.randint(0,25)), '_id': i*1000+j}
+                if element['key'] not in expected_count:
+                    expected_count[element['key']]=1
+                else:
+                    expected_count[element['key']] += 1
+                list_of_elements.append(element)
+            self.channel.basic_publish(exchange='', routing_key='pipeline_start',
+                                       body=json.dumps(list_of_elements))
+        self.channel.basic_publish(exchange='', routing_key='pipeline_start',
+                                   body=json.dumps({}))
         consume_process = Process(target=self._read_process, args=(self.write_pipe,
                                                                    'pipelineC_result'))
         consume_process.start()
-        for element in MESSAGES_FOR_TESTING_ENDING_2:
-            self.channel.basic_publish(exchange='', routing_key='pipeline_start',
-                                       body=json.dumps(element))
         consume_process.join()
         processed_data = []
         while not processed_data or processed_data[-1] != WINDOW_END_MESSAGE:
@@ -237,7 +235,8 @@ class TestIntegrations(unittest.TestCase):
                     count_result[item['key']] = item['count']
             else:
                 count_result[resp['key']] = resp['count']
-        self.assertEqual(TESTING_RESULT, count_result)
+        self.assertEqual({k:v for k,v in  expected_count.items() if v>2}, count_result)
+        chaos_monkey_p.terminate()
 
     def setUp(self) -> None:
         try:
@@ -250,12 +249,19 @@ class TestIntegrations(unittest.TestCase):
         self.queues_to_purge = []
         self.dirs_to_delete = []
         self.recv_pipe, self.write_pipe = Pipe(False)
+        self.recv_chaos, self.write_chaos = Pipe(False)
         self.connection = pika.BlockingConnection(pika.ConnectionParameters(host="localhost"))
         self.channel = self.connection.channel()
         self._setup_queue('pipeline_start')
 
     def tearDown(self) -> None:
+        while self.recv_chaos.poll(0.1):
+            pid=self.recv_chaos.recv()
+            try:
+                os.kill(pid, signal.SIGKILL)
+            except Exception:
+                continue
         for p in self.processes_to_join:
-            p.terminate()
+            p[0].terminate()
         for l in self.dirs_to_delete:
             shutil.rmtree(l, ignore_errors=True)
